Scala getting started working locally and on MSF (downgraded to Scala 2.12)
