+++ scala/GettingStarted/README.md
@@ -3,7 +3,7 @@
 Skeleton project for a basic Flink Java application to run on Amazon Managed Service for Apache Flink.
 
 * Flink version: 1.20
-* Scala version: 3.3.0
+* Scala version: 2.12.20
 * Flink API: DataStream API
 * Language: Java (11)
 
@@ -12,11 +12,13 @@ The project can run both on Amazon Managed Service for Apache Flink, and locally
 The application shows how to get runtime configuration, and sets up a Kinesis Data Stream source and a sink.
 
 ### Pre-requisites
+
 You need to have `sbt` tool installed on you machine to build a Scala project. Use [steps from official guide](https://www.scala-sbt.org/download.html) to do that.
 
 ### Build
+
 - Run `sbt assembly` to build an uber jar
-- Use `target/scala-3.3.0/ScalaGettingStarted-flink_1.20.jar` in your MSF application
+- Use `target/scala-2.12/ScalaGettingStarted-flink_1.20.jar` in your MSF application
 
 ### Runtime configuration
 
@@ -25,12 +27,19 @@ Apache Flink.
 
 The following runtime properties are required:
 
-| Group ID       | Key           | Description              |
-|----------------|---------------|--------------------------|
-| `InputStream0` | `stream.name` | Name of the input stream |
-| `InputStream0` | `aws.region`  | Region of the input stream, e.g. `us-east-1` |
+| Group ID        | Key           | Description                                   |
+|-----------------|---------------|-----------------------------------------------|
+| `InputStream0`  | `stream.name` | Name of the input stream                      |
+| `InputStream0`  | `aws.region`  | Region of the input stream, e.g. `us-east-1`  |
+| `OutoutStream0` | `stream.name` | Name of the output stream                     |
+| `OutoutStream0` | `aws.region`  | Region of the output stream, e.g. `us-east-1` |
 
-Runtime Properties are expected in the Group IDs `InputStream0` and `OutputStream0`.
+
+### Running in IntelliJ
+
+You can run this example directly in IntelliJ, without any local Flink cluster or local Flink installation.
+
+See [Running examples locally](../../java/running-examples-locally.md) for details.
 
 ### Generating data
